Separated RxJava1 and 2 Answers
